[seller] refactor(204):개별 이미지 프리뷰 컴포넌트를 분리
