+++ apps/seller/src/features/products/create/create-form-thumbnail/create-form-thumbnail-area.ui.tsx
@@ -1,5 +1,3 @@
-import React, { useEffect } from 'react'
-
 import {
   Button,
   Dialog,
@@ -10,25 +8,14 @@ import {
   DialogTitle,
   Label,
 } from '@dessert/ui'
-import {
-  DndContext,
-  DragEndEvent,
-  KeyboardSensor,
-  PointerSensor,
-  closestCenter,
-  useSensor,
-  useSensors,
-} from '@dnd-kit/core'
-import {
-  SortableContext,
-  arrayMove,
-  rectSortingStrategy,
-  sortableKeyboardCoordinates,
-  useSortable,
-} from '@dnd-kit/sortable'
-import { CSS } from '@dnd-kit/utilities'
-import { Camera, XIcon } from 'lucide-react'
+import { DndContext, closestCenter } from '@dnd-kit/core'
+import { SortableContext, rectSortingStrategy } from '@dnd-kit/sortable'
+import { Camera } from 'lucide-react'
 
+import {
+  ImagePreviewItem,
+  SortableImageItems,
+} from './create-form-thumbnail-items.ui'
 import { useProductThumbnailForm } from './use-product-thumbnail.hook'
 // import { useCreateHeaderSteps } from '../create-store'
 
@@ -123,7 +110,7 @@ export const ThumbnailUploadArea = () => {
               strategy={rectSortingStrategy}
             >
               {(extraImages as File[]).map((file, idx) => (
-                <SortableImageItem
+                <SortableImageItems
                   key={`${file.name}-${file.lastModified}`}
                   id={`${file.name}-${file.lastModified}`}
                   file={file}
@@ -181,45 +168,6 @@ function UploadButton({
   )
 }
 
-function ImagePreviewItem({
-  file,
-  onDelete,
-}: {
-  file: File
-  onDelete: () => void
-}) {
-  const [previewUrl, setPreviewUrl] = React.useState<string>('')
-
-  useEffect(() => {
-    // URL 생성
-    const url = URL.createObjectURL(file)
-    setPreviewUrl(url)
-
-    // 클린업 함수: 컴포넌트가 언마운트되거나 file이 바뀔 때 메모리를 해제합니다.
-    return () => {
-      URL.revokeObjectURL(url)
-    }
-  }, [file])
-  return (
-    <div className="relative h-[120px] w-[120px] overflow-hidden rounded-16 border border-gray-100 bg-gray-50">
-      {previewUrl && (
-        <img
-          src={previewUrl}
-          alt="preview"
-          className="size-full object-cover"
-        />
-      )}
-      <button
-        type="button"
-        onClick={onDelete}
-        className="absolute top-6 right-6 rounded-full bg-black/10 p-2 text-white transition-colors hover:bg-gray-900/50"
-      >
-        <XIcon className="size-12" />
-      </button>
-    </div>
-  )
-}
-
 interface DeleteConfirmDialogProps {
   isOpen: boolean
   onClose: () => void
@@ -257,43 +205,5 @@ function DeleteConfirmDialog({
         </DialogFooter>
       </DialogContent>
     </Dialog>
-  )
-}
-function SortableImageItem({
-  id,
-  file,
-  onDelete,
-}: {
-  id: string
-  file: File
-  onDelete: () => void
-}) {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id })
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-    zIndex: isDragging ? 50 : 1,
-    opacity: isDragging ? 0.5 : 1,
-  }
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      {...attributes}
-      {...listeners}
-      // 드래그 중 커서 모양 변경 및 터치 스크롤 방지
-      className="cursor-grab touch-none active:cursor-grabbing"
-    >
-      <ImagePreviewItem file={file} onDelete={onDelete} />
-    </div>
   )
 }
