+++ apps/seller/src/features/products/create/create-form-thumbnail/index.ts
@@ -2,3 +2,7 @@ export { thumbnailSchema } from './create-form-thumbnail.schema'
 export { ThumbnailUploadArea } from './create-form-thumbnail-area.ui'
 export { useProductThumbnailForm } from './use-product-thumbnail.hook'
 export type { ThumbnailFormType } from './create-form-thumbnail.type'
+export {
+  ImagePreviewItem,
+  SortableImageItems,
+} from './create-form-thumbnail-items.ui'
